refactor(hello-work-job-searcher): コンポーネント構成・ディレクトリ構造の整理
- components配下を_components配下へリネーム・整理
- atom関連ファイルの新設・移動
- importパスの修正
- 旧components/client/atom/index.tsの削除
- ページ・ルート等のimportパスを新構成に対応
