+++ apps/hello-work-job-searcher/src/app/_components/Job.tsx
