+++ apps/hello-work-job-searcher/src/app/_store/type.ts
