+++ apps/hello-work-job-searcher/src/app/jobs/loading.tsx
@@ -1,4 +1,4 @@
-import Loading from "../components/Loading";
+import Loading from "../_components/Loading";
 
 export default function Page() {
   // Or a custom loading skeleton component
